refactor(backend): move location_service and repository_service to app/services
